# openapi.yaml
openapi: 3.1.0
info:
  title: Scrutineer Skill API
  version: 0.4.0
  description: |
    A narrow HTTP surface that claude-code skills use while they run. The
    goal is enough to let a skill orchestrate other skills and read prior
    scan outputs without pushing that logic into scrutineer's Go code.

    Every scrutineer skill scan is issued a bearer token that is staged into
    the workspace context.json. Skills authenticate with that token. A token
    is only valid while its originating scan is running, and only grants
    access to the repository the scan was issued against.

    The `/v1/*` paths are an unauthenticated, localhost-only export surface
    intended for external scripts (jq, cron, CI dumps). They emit JSONL and
    rely on the same host-header check as the browser UI; see threatmodel.md.
servers:
  - url: http://127.0.0.1:8080/api
    description: Local scrutineer server
security:
  - bearerAuth: []
paths:
...
  /repositories/{id}/findings:
    get:
      summary: List findings across every scan on this repository
      operationId: listRepositoryFindings
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
        - name: status
          in: query
          schema:
            type: string
            enum: [new, enriched, triaged, ready, reported, acknowledged, fixed, published, rejected, duplicate]
        - name: skill
          in: query
          description: Only return findings produced by scans of this skill (e.g. `semgrep`, `security-deep-dive`).
          schema: { type: string }
        - name: scan_group
          in: query
          description: Only return findings from scans in this parallel batch, so an in-flight audit skill reads only its siblings' findings.
          schema: { type: string }
        - name: sub_path
          in: query
          description: Only return findings for this monorepo sub-package (e.g. `activesupport`), so a repository can be exported or bundled one sub-package at a time.
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingSummary' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Stream one finding into the concurrent-finding log
      description: >
        Records a single finding emitted mid-scan so sibling scans in the same
        `scan_group` can read it before this scan completes. The body is one
        finding in the same shape it has in the skill's `report.json`; the
        scan's identity (scan id, repository, commit, sub-path) is stamped from
        the bearer token, never trusted from the body. The finding still
        belongs in the final `report.json`, which reconciles against the
        streamed row by fingerprint rather than duplicating it.
      operationId: streamRepositoryFinding
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FindingInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /findings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: Fetch one finding with its six-step prose
      operationId: getFinding
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update mutable fields on a finding
      operationId: patchFinding
      description: |
        Each field value is written through the history-tracked path.
        Skills write with source=model_suggested; the browser UI (behind
        its own session) writes with source=analyst. Unknown field names
        are rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fields]
              properties:
                by:
                  type: string
                  description: Optional author string for the history row.
                fields:
                  type: object
                  additionalProperties: { type: string }
                  description: |
                    Map of field name to new value. Allowed fields:
                    title, severity, status, cwe, location, affected, reachability, quality_tier,
                    cve_id, ghsa_id, cvss_vector, cvss_v4_vector, fix_version, fix_commit,
                    resolution, disclosure_draft, suggested_recipients, assignee,
                    suggested_fix, suggested_fix_commit, breaking_change, breaking_change_rationale,
                    exploited_in_wild, exploited_in_wild_evidence, mitigation, mitigation_semgrep,
                    release_tag, release_url, last_revalidate_verdict.
      responses:
        '204': { description: Updated }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /findings/{id}/communications:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: List logged communications for a finding
      operationId: listFindingCommunications
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingCommunication' }
    post:
      summary: Log a communication (sent or received) about a finding
      operationId: addFindingCommunication
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FindingCommunication' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingCommunication' }
  /findings/{id}/references:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: List references attached to a finding
      operationId: listFindingReferences
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingReference' }
    post:
      summary: Attach an external reference to a finding
      operationId: addFindingReference
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FindingReference' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingReference' }
...
  /v1/findings:
    get:
      summary: Export findings across all repositories as JSONL
      description: |
        Returns one JSON object per line (`application/x-ndjson`). No bearer
        token; localhost-only. Each line is a `FindingExport` carrying the
        full set of `db.Finding` columns.
      operationId: exportFindings
      security: []
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [jsonl] }
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
        - name: status
          in: query
          schema:
            type: string
            enum: [new, enriched, triaged, ready, reported, acknowledged, fixed, published, rejected, duplicate]
      responses:
        '200':
          description: JSONL stream of FindingExport objects, one per line.
          content:
            application/x-ndjson:
              schema: { $ref: '#/components/schemas/FindingExport' }
        '400':
          description: Unsupported format query value, or `encrypt`/`scope`/`include` requested (only valid on per-repository bundle exports).
        '403':
          description: Host header is not localhost
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RepositoryID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Token does not grant access to this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Malformed query or request body, or a missing required field
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    FindingInput:
      type: object
      description: >
        One finding as it appears in a skill's report.json, streamed into the
        concurrent-finding log. Only title, severity and location are required;
        the other report.json finding fields (cwe, sinks, dup_check, trace, …)
        are accepted and stored. Scan-owned identity fields are ignored.
      required: [title, severity, location]
      additionalProperties: true
      properties:
        title:    { type: string }
        severity: { type: string, enum: [Critical, High, Medium, Low] }
        location: { type: string }
        cwe:      { type: string }
        sinks:    { type: array, items: { type: string } }
        dup_check: { type: string, description: "One sentence: which siblings under this scan_group you compared against and why this one is distinct." }
...
    FindingDetail:
      allOf:
        - $ref: '#/components/schemas/FindingSummary'
        - type: object
          description: Adds the six-step narrative and the disclosure draft/routing fields.
          properties:
            trace:            { type: string }
            boundary:         { type: string }
            validation:       { type: string }
            prior_art:        { type: string }
            reach:            { type: string }
            rating:           { type: string }
            disclosure_draft: { type: string }
            suggested_recipients: { type: string, description: "File-level owners for the finding's location, from CODEOWNERS or recent non-bot committers." }
            sub_path:         { type: string, description: "Monorepo sub-folder the producing scan was scoped to. The finding's location is relative to it; empty means the repo root." }
            verification:
              allOf:
                - $ref: '#/components/schemas/FindingVerification'
              nullable: true
            remediation:
              allOf:
                - $ref: '#/components/schemas/RemediationAttempt'
              nullable: true
...
    FindingReference:
      type: object
      required: [url]
      properties:
        id:         { type: integer, format: int64 }
        finding_id: { type: integer, format: int64 }
        url:        { type: string, format: uri }
        tags:       { type: string, description: Comma-joined tags (issue, pr, cve, ghsa, patch, advisory, discussion, article). }
        summary:    { type: string }
        created_at: { type: string, format: date-time }
...